#356 Stabilize Smart Event code
There appear to be instances when the smart events aren't
updating everything they should be. Before this change,
the code would only add events to the smart list if the
exact same arguments weren't already in there.

This change makes it so it will always add things to the
smart list, but only run each event once per frame. This
would mean if there is a large delay, it may be a delay,
but it would have to be quite long. Otherwise, it should
sort itself out quickly.